# Optimize `offerRecordsToStreams` code