Before:
```scala

          ns  percent  samples  top
  ----------  -------  -------  ---
 12150000000   47.13%     1215  Unique_Node_List::remove(Node*)
  2340000000    9.08%      234  __psynch_cvwait
   510000000    1.98%       51  itable stub
   350000000    1.36%       35  org.apache.kafka.clients.consumer.internals.Fetcher.maybeLeaderEpoch
   320000000    1.24%       32  zio.kafka.consumer.internal.Runloop.$anonfun$offerRecordsToStreams$1
   310000000    1.20%       31  __psynch_cvsignal
   200000000    0.78%       20  kevent
   170000000    0.66%       17  scala.runtime.BoxesRunTime.boxToBoolean
   150000000    0.58%       15  org.apache.kafka.clients.consumer.internals.Fetcher$CompletedFetch.nextFetchedRecord
   140000000    0.54%       14  _platform_bzero$VARIANT$Haswell
   140000000    0.54%       14  sendfile
```

After:
```scala

          ns  percent  samples  top
  ----------  -------  -------  ---
  1240000000   15.80%      124  __psynch_cvwait
   380000000    4.84%       38  org.apache.kafka.clients.consumer.internals.Fetcher.maybeLeaderEpoch
   270000000    3.44%       27  __psynch_cvsignal
   230000000    2.93%       23  G1ParScanThreadState::do_copy_to_survivor_space(G1HeapRegionAttr, oopDesc*, markWord)
   180000000    2.29%       18  itable stub
   180000000    2.29%       18  void G1ParScanThreadState::do_oop_evac<narrowOop>(narrowOop*)
   140000000    1.78%       14  java.nio.HeapByteBuffer.get
   120000000    1.53%       12  jbyte_disjoint_arraycopy
   120000000    1.53%       12  org.apache.kafka.clients.consumer.internals.Fetcher$CompletedFetch.fetchRecords
   120000000    1.53%       12  org.apache.kafka.common.utils.ByteUtils.readUnsignedVarint
   110000000    1.40%       11  org.apache.kafka.clients.consumer.internals.Fetcher$CompletedFetch.nextFetchedRecord
   110000000    1.40%       11  sendfile
   110000000    1.40%       11  org.apache.kafka.common.record.DefaultRecordBatch.partitionLeaderEpoch
    80000000    1.02%        8  java.nio.Buffer.<init>
    80000000    1.02%        8  scala.runtime.BoxesRunTime.boxToBoolean
    80000000    1.02%        8  org.apache.kafka.clients.consumer.internals.Fetcher.parseRecord
    70000000    0.89%        7  zio.ChunkBuilder.<init>
    70000000    0.89%        7  jdk.internal.misc.Unsafe.park
    60000000    0.76%        6  kevent
    60000000    0.76%        6  org.apache.kafka.common.record.DefaultRecord.readFrom
    60000000    0.76%        6  G1ParScanThreadState::trim_queue_to_threshold(unsigned int)
    60000000    0.76%        6  updateBytesCRC32C
    60000000    0.76%        6  void G1ScanEvacuatedObjClosure::do_oop_work<narrowOop>(narrowOop*)
    50000000    0.64%        5  _platform_bzero$VARIANT$Haswell
    50000000    0.64%        5  I2C/C2I adapters
    50000000    0.64%        5  java.nio.Buffer.nextGetIndex
    40000000    0.51%        4  org.apache.kafka.clients.consumer.internals.Fetcher.access$3400
    40000000    0.51%        4  read
    40000000    0.51%        4  CodeHeap::find_blob_unsafe(void*) const
    30000000    0.38%        3  thread_self_trap
    30000000    0.38%        3  PhaseIdealLoop::split_if_with_blocks(VectorSet&, Node_Stack&)
    30000000    0.38%        3  __gettimeofday
    30000000    0.38%        3  zio.Fiber$Status$Running.<init>
    30000000    0.38%        3  java.nio.Buffer.limit
    30000000    0.38%        3  zio.ZIO$$$Lambda$61.0x0000000800cea5d0.apply
    30000000    0.38%        3  PhaseCCP::analyze()
    30000000    0.38%        3  IndexSetIterator::advance_and_next()
    30000000    0.38%        3  PhaseChaitin::build_ifg_physical(ResourceArea*)
    30000000    0.38%        3  PhaseIdealLoop::build_loop_late_post_work(Node*, bool)
    30000000    0.38%        3  __psynch_mutexwait
    30000000    0.38%        3  Interpreter
    30000000    0.38%        3  jint_disjoint_arraycopy
    30000000    0.38%        3  zio.kafka.consumer.internal.Runloop.$anonfun$offerRecordsToStreams$5
    30000000    0.38%        3  java.lang.invoke.DirectMethodHandle.allocateInstance
    20000000    0.25%        2  zio.stream.internal.ChannelExecutor$.read$1
```